Middleware for Storage
Added diskStorage and addDirectory as middleware in fileStorage.js
